% Generated by roxygen2 (4.0.1): do not edit by hand
\name{countbot}
\alias{countbot}
\title{Function Description}
\usage{
countbot(formula, data, link = c("logit", "probit", "cloglog", "cauchit"),
  family = c("poisson", "negative binomial"), subset,
  na.action = getOption("na.action"), fisher = FALSE, conf.int = 0.95)
}
\arguments{
\item{formula}{}

\item{data}{}

\item{link}{}

\item{family}{}

\item{subset}{an optional vector specifying a subset of
observations to be used in the fitting process.}

\item{na.action}{action to take when \code{NA} present in data.}

\item{fisher}{logical indicating whether to apply fisher transformation to compute confidence intervals and p-values for the correlation.}

\item{conf.int}{numeric specifying confidence interval coverage.}
}
\value{
object of \samp{countbot} class.
}
\description{
What the function does
}
\details{
Details of function
}
\examples{
data(PResidData)
cocobot(y|w ~ z, data=PResidData)
}

