% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.conditional.Spearman.R
\name{plot.conditional.Spearman}
\alias{plot.conditional.Spearman}
\title{conditional.Spearman class plot method}
\usage{
\method{plot}{conditional.Spearman}(x, ...)
}
\arguments{
\item{x}{conditional.Spearman object}

\item{...}{arguments passed to plot.default}
}
\description{
conditional.Spearman class plot method
}
\keyword{plot}
