% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.conditional_Spearman.R
\name{plot.conditional_Spearman}
\alias{plot.conditional_Spearman}
\title{conditional_Spearman class plot method}
\usage{
\method{plot}{conditional_Spearman}(x, ...)
}
\arguments{
\item{x}{conditional_Spearman object}

\item{...}{arguments passed to plot.default}
}
\description{
conditional_Spearman class plot method
}
\keyword{plot}
