% Generated by roxygen2 (4.0.1): do not edit by hand
\name{presid}
\alias{presid}
\title{Probablity Scale Residual}
\usage{
presid(mod, ...)
}
\arguments{
\item{mod}{The model object for which the probality scale residual is calculated}

\item{...}{Aditional arguements passed to methods}
}
\value{
The probablity scale residual for the model
}
\description{
\code{presid} calculates the probablity scale residual for various model types
}
\examples{
library(survival)
library(actuar) #for getting the cdf of loglogistic cdf

set.seed(100)
time <- sample(1:60, size=100, replace=TRUE)
delta <- sample(c(0,1), size=100, prob=c(0.2,0.8), replace=TRUE)
X <- round(rnorm(100, mean=36, sd=7), 0)

mod <- survreg(Surv(time, delta) ~ X, dist="weibull")
summary(presid.survreg(mod))

#Example for coxph
mod <- coxph(Surv(time, delta) ~ X)
summary(presid.coxph(mod))
##################


#Example for negative binomial
library(MASS)
n <- 2000
beta0 <- 0
beta1 <- 1
X <- rnorm(n)
mu <- beta0 + beta1*X
phi <- 3
y <- rnbinom(n, mu=exp(mu), size=phi)

mod <- glm.nb(y ~ X)
summary(presid.nb(mod))
}
\author{
Charles Dupont \email{charles.dupont@vanderbilt.edu}

Chun Li \email{chun.li@vanderbilt.edu}

Bryan Shepherd \email{bryan.shepherd@vanderbilt.edu}

Valintine Wanga \email{?}
}
\references{
Li C and Shepherd BE, A new residual for ordinal
outcomes. Biometrika 2012; 99:473-480
}

