% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PResiduals-package.R
\docType{package}
\name{PResiduals-package}
\alias{PResiduals-package}
\alias{PResiduals}
\title{Computes probability-scale residuals and residual correlations.}
\description{
This package outputs probability-scale residuals from multiple
models and computes residual correlation.  Probability-scale
residual can be computed for continuous, ordinal, binary, count,
and time-to-event data (although the current implementation is only
for ordinal variables). Plots of probability-scale residuals can be
useful for model diagnostics.  Residual correlation can be used to
test for conditional independence between multiple types of
variables.
}
\author{
Bryan Shepherd \email{bryan.shepherd@vanderbilt.edu}

Chun Li \email{cxl791@case.edu}

Qi Liu \email{qi.liu4@merck.com}

Charles Dupont \email{charles.dupont@vanderbilt.edu}

Jeffrey Horner \email{jeffrey.horner@vanderbilt.edu}
}
\keyword{package}
