% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cobot.R
\name{print.cobot}
\alias{print.cobot}
\title{cobot class print method}
\usage{
\method{print}{cobot}(x, ...)
}
\arguments{
\item{x}{cobot object}

\item{...}{arguments passed to print.default}
}
\description{
cobot class print method
}
\keyword{print}
