% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cocobot.R
\name{print.cocobot}
\alias{print.cocobot}
\title{cocobot class print method}
\usage{
\method{print}{cocobot}(x, ...)
}
\arguments{
\item{x}{cocobot object}

\item{...}{arguments passed to print.default}
}
\description{
cocobot class print method
}
\keyword{print}
