% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.conditional.Spearman.R
\name{print.conditional.Spearman}
\alias{print.conditional.Spearman}
\title{conditional.Spearman class print method}
\usage{
\method{print}{conditional.Spearman}(x, ...)
}
\arguments{
\item{x}{conditional.Spearman object}

\item{...}{arguments passed to print.default}
}
\description{
conditional.Spearman class print method
}
\keyword{print}
