% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.conditional_Spearman.R
\name{print.conditional_Spearman}
\alias{print.conditional_Spearman}
\title{conditional_Spearman class print method}
\usage{
\method{print}{conditional_Spearman}(x, ...)
}
\arguments{
\item{x}{conditional_Spearman object}

\item{...}{arguments passed to print.default}
}
\description{
conditional_Spearman class print method
}
\keyword{print}
