% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.partial.Spearman.R
\name{print.partial.Spearman}
\alias{print.partial.Spearman}
\title{partial.Spearman class print method}
\usage{
\method{print}{partial.Spearman}(x, ...)
}
\arguments{
\item{x}{partial.Spearman object}

\item{...}{arguments passed to print.default}
}
\description{
partial.Spearman class print method
}
\keyword{print}

