% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.partial_Spearman.R
\name{print.partial_Spearman}
\alias{print.partial_Spearman}
\title{partial_Spearman class print method}
\usage{
\method{print}{partial_Spearman}(x, ...)
}
\arguments{
\item{x}{partial_Spearman object}

\item{...}{arguments passed to print.default}
}
\description{
partial_Spearman class print method
}
\keyword{print}
