% Generated by roxygen2 (4.0.1): do not edit by hand
\name{countbot}
\alias{countbot}
\title{Conditional count by ordinal tests for association.}
\usage{
countbot(formula, data, link.x = c("logit", "probit", "cloglog", "cauchit"),
  family.y = c("poisson", "negative binomial"), subset,
  na.action = getOption("na.action"), fisher = FALSE, conf.int = 0.95)
}
\arguments{
\item{formula}{an object of class \code{\link{Formula}} (or one
that can be coerced to that class): a symbolic description of the
model to be fitted.  The details of model specification are given
under \sQuote{Details}.}

\item{data}{an optional data frame, list or environment (or object
coercible by \code{\link{as.data.frame}} to a data frame)
containing the variables in the model.  If not found in
\code{data}, the variables are taken from
\code{environment(formula)}, typically the environment from which
\code{countbot} is called.}

\item{link.x}{The link family to be used for the ordinal model of
\var{X} on \var{Z}.  Defaults to \samp{logit}. Other options are
\samp{probit}, \samp{cloglog}, and \samp{cauchit}.}

\item{family.y}{The error distribution for the count model of \var{Y} on \var{Z}.
Defaults to \samp{poisson}. The other option is \samp{negative binomial}.
If \samp{negative binomial} is specified, \code{\link[MASS]{glm.nb}} is called to fit the count model.}

\item{subset}{an optional vector specifying a subset of
observations to be used in the fitting process.}

\item{na.action}{action to take when \code{NA} present in data.}

\item{fisher}{logical indicating whether to apply fisher transformation to compute confidence intervals and p-values for the correlation.}

\item{conf.int}{numeric specifying confidence interval coverage.}
}
\value{
object of \samp{countbot} class.
}
\description{
\code{countbot} tests for independence between an ordered categorical
variable, \var{X}, and a count variable, \var{Y}, conditional on other variables,
\var{Z}.  The basic approach involves fitting an ordinal model of \var{X} on
\var{Z}, a Poisson or Negative Binomial model of \var{Y} on \var{Z}, and then determining whether there is any
residual information between \var{X} and \var{Y}.  This is done by
computing residuals for both models, calculating their correlation, and
testing the null of no residual correlation.  This procedure is analogous to test statistic
\code{T2} in \code{cobot}.  Two test statistics (correlations) are currently output.  The first
is the correlation between probability-scale residuals. The second is the correlation between
the Pearson residual for the count outcome model and a latent variable residual
for the ordinal model.
}
\details{
Formula is specified as \code{\var{X} | \var{Y} ~ \var{Z}}.
This indicates that models of \code{\var{X} ~ \var{Z}} and
\code{\var{Y} ~ \var{Z}} will be fit.  The null hypothesis to be
tested is \eqn{H_0 : X}{H0 : X} independent of \var{Y} conditional
on \var{Z}.  The ordinal variable, \code{\var{X}}, must precede the \code{|} and be a factor variable, and \code{\var{Y}} must be an integer.
}
\examples{
generate.data3 = function(alphax, betax, alphay, betay, eta, N) {
  z = rnorm(N,0,1)
  x = v = numeric(N)

  ## px is an N x length(alphax) matrix.
  ## Each row has the TRUE cummulative probabilities for each subject.
  px = (1 + exp(- outer(alphax, betax*z, "+"))) ^ (-1)
  aa = runif(N)
  for(i in 1:N)
    x[i] = sum(aa[i] > px[,i])
  x = as.numeric(as.factor(x))
  v = rpois(N, exp(outer(alphay, betay*z+eta[x], "+")))

  return(list(x=as.factor(x), v=v, z=z))
}

set.seed(13)
alphax = c(-1, 0, 1, 2)
betax = 1
alphay = 1
betay = -.5

#eta = rep(0, 5)
eta = c(1:5)/5
N = 100

data <- generate.data3(alphax, betax, alphay, betay, eta, N)
###
countbot(x|v~z, data=data, link.x="logit", family.y="poisson")
countbot(x|v~z, data=data, link.x="probit", family.y="negative binomial")
}
\references{
Li C and Shepherd BE (2012)
A new residual for ordinal outcomes.
\emph{Biometrika}. \bold{99}: 473--480.

Shepherd BE, Li C, Liu Q (submitted)
Probability-scale residuals for continuous, discrete, and censored data.
}

