% Generated by roxygen2 (4.0.1): do not edit by hand
\name{presid}
\alias{presid}
\title{Probability-scale Residual}
\usage{
presid(object, ...)
}
\arguments{
\item{object}{The model object for which the probability-scale residual is calculated}

\item{...}{Additional arguements passed to methods}
}
\value{
The probability scale residual for the model
}
\description{
\code{presid} Calculates the probability-scale residual for various model
function objects. Currently supported models include \code{\link{glm}}
(Poisson, binomial, and gaussian families), \code{\link{lm}} in the
\pkg{stats} library, \code{\link{survreg}} (Weibull, exponential, gaussian,
logistic, and lognormal distributions) and \code{\link{coxph}} in the
\pkg{survival} library, and \code{\link{polr}} and \code{\link{glm.nb}} in
the \pkg{MASS} library.
}
\details{
Probability-scale residual is \eqn{P(Y* < y) - P(Y* > y)} where \eqn{y} is the observed
outcome and \eqn{Y*} is a random variable from the fitted distribution.
}
\examples{
library(survival)
library(stats)

set.seed(100)
n <- 1000
x <- rnorm(n)
t <- rweibull(n, shape=1/3, scale=exp(x))
c <- rexp(n, 1/3)
y <- pmin(t, c)
d <- ifelse(t<=c, 1, 0)

mod.survreg <- survreg(Surv(y, d) ~ x, dist="weibull")
summary(presid(mod.survreg))
plot(x, presid(mod.survreg))

#Example for coxph
time <- sample(60, size=n, replace=TRUE)
delta <- sample(c(0, 1), size=n, prob=c(0.2, 0.8), replace=TRUE)
X <- round(rnorm(100, mean=36, sd=7), 0)

mod.coxph <- coxph(Surv(time, delta) ~ X)
summary(presid(mod.coxph))

#Example for negative binomial
library(MASS)
n <- 2000
beta0 <- 0
beta1 <- 1
X <- rnorm(n)
mu <- beta0 + beta1*X
phi <- 3
y <- rnbinom(n, mu=exp(mu), size=phi)

mod.glm.nb <- glm.nb(y ~ X)
summary(presid(mod.glm.nb))
}
\author{
Charles Dupont \email{charles.dupont@vanderbilt.edu}

Chun Li \email{chun.li@vanderbilt.edu}

Bryan Shepherd \email{bryan.shepherd@vanderbilt.edu}
}
\references{
Shepherd BE, Li C, Liu Q.  Probability-scale residuals for continuous,
discrete, and censored data.  Submitted.

Li C and Shepherd BE, A new residual for ordinal
outcomes. Biometrika 2012; 99:473-480
}

