\name{PResidData}
\alias{PResidData}
\docType{data}
\title{
PResidual Example Dataset
}
\description{
An example dataset for use with PResidual
}
\usage{data(PResidData)}
\format{
  A data frame with 500 observations on the following 5 variables.
  \describe{
    \item{\code{x}}{an ordered factor with levels \code{1} < \code{2} < \code{3} < \code{4} < \code{5}}
    \item{\code{y}}{an ordered factor with levels \code{1} < \code{2} < \code{3} < \code{4}}
    \item{\code{z}}{a numeric vector}
    \item{\code{w}}{a numeric vector}
    \item{\code{c}}{a numeric vector of non-negative integers}
  }
}
\examples{
data(PResidData)
}
\keyword{datasets}
