% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PResidData.R
\docType{data}
\name{PResidData}
\alias{PResidData}
\title{Example Dataset for PResiduals Package}
\format{
A data frame with 200 rows and 5 variables:
\describe{
  \item{x}{an ordered categorical variable with 5 levels}
  \item{y}{an ordered categorical variable with 4 levels}
  \item{z}{a continuous variable}
  \item{w}{a continuous variable}
  \item{c}{a count variable}
}
}
\source{
Simulated
}
\usage{
PResidData
}
\description{
This is a dataset used in Examples Section of PResiduals package help files.
}
\keyword{datasets}
